#13 refactor: realocando o controlador de preferência e de interesses do usuário e também os serviços de ambos | fix: corringdo retorno de erros
