+++ backend/src/routes/user.routes.ts
@@ -1,6 +1,7 @@
 import { Router } from 'express';
 import { UserController } from '../controllers/user.controller';
-
+import { UserPreferencesController } from '../controllers/userPreferences.controller';
+import { TravelInterestsController } from '../controllers/travelInterests.controller';
 const router = Router();
 
 /**
@@ -210,11 +211,11 @@ router.put('/:id', (req, res) => UserController.updateUser(req, res));
  */
 router.delete('/:id', (req, res) => UserController.deleteUser(req, res));
 
-router.get('/:id/preferences', UserController.getUserTravelPreferencesByUserId);
-router.post('/:id/preferences', UserController.saveUserTravelPreferences);
-router.put('/:id/preferences', UserController.saveUserTravelPreferences);
+router.get('/:id/preferences', UserPreferencesController.getUserTravelPreferencesByUserId);
+router.post('/:id/preferences', UserPreferencesController.saveUserTravelPreferences);
+router.put('/:id/preferences', UserPreferencesController.saveUserTravelPreferences);
 
-router.get('/:id/interests', UserController.getUserTravelInterestsByUserId);
-router.post('/:id/interests', UserController.saveUserTravelInterests);
-router.put('/:id/interests', UserController.saveUserTravelInterests);
+router.get('/:id/interests', TravelInterestsController.getUserTravelInterestsByUserId);
+router.post('/:id/interests', TravelInterestsController.saveUserTravelInterests);
+router.put('/:id/interests', TravelInterestsController.saveUserTravelInterests);
 export default router;
